updated data generation

+++ database/src/main/java/com/seitenbau/testing/dbunit/generator/values/ValueGenerator.java
@@ -2,7 +2,7 @@ package com.seitenbau.testing.dbunit.generator.values;
 
 public interface ValueGenerator
 {
-  void initialize(int seed);
+  void initialize(long seed);
 
   String nextValue();
 }
